requirements.txt
-e .
streamlit>=1.32.0
langchain>=0.1.0
langchain-openai>=0.0.2,<0.0.3
openai>=1.12.0
python-dotenv>=1.0.0
pandas>=2.2.0
plotly>=5.18.0
geopy>=2.4.1
tiktoken==0.5.2
crewai>=0.11.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation-crewai>=0.1.0
# opentelemetry-instrumentation-streamlit>=0.1.0
arize-phoenix-otel
openinference-instrumentation-crewai
crewai-tools
openinference-instrumentation-litellm
